requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-cognitiveservices-speech==1.43.0
boto3==1.38.3
botocore==1.38.3
cachetools==5.5.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
elevenlabs==1.57.0
filelock==3.18.0
frozenlist==1.6.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
inflection==0.5.1
jiter==0.9.0
jmespath==1.0.1
multidict==6.4.3
openai==1.76.0
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
pyht==0.1.14
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
resemble==1.5.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
speechify==0.0.0
speechify-api==1.2.0
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.1
yarl==1.20.0
whitenoise==6.6.0
